refactor(common): move `about.interface.ts` from `customTypes` to `area/types/area.ts`
