+++ apps/frontend/src/lib/components/dashboard/area/AreaCreationDialog/AreaCreationDialog.svelte
@@ -3,6 +3,7 @@
 <script lang="ts">
     import { onMount, afterUpdate } from "svelte";
     import type { ActionData } from "../../../../../routes/[lang=lang]/dashboard/$types";
+    import type { Services, Action, Reaction } from "@common/area/types/area";
     import {  buttonVariants } from "$lib/components/ui/button";
     import * as Dialog from "$lib/components/ui/dialog";
     import Combobox from "$lib/components/dashboard/area/Combobox/Combobox.svelte";
